src/Plugins/Input/Database/requirements.txt
psycopg2-binary>=2.9.9  # PostgreSQL driver
mysql-connector-python>=8.2.0  # MySQL/MariaDB driver
pymongo>=4.6.0  # MongoDB driver